requirements.txt
requests
Flask
google-generative-ai
textblob
pillow
pymongo